+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.7.1] - 2023-04-05
+
 ## [3.7.0] - 2023-04-05
 ### Added
 - German translation.